refactor core stucture
