+++ packages/core/src/configs/buildSchemes.ts
