+++ packages/core/src/configs/index.ts
@@ -1,206 +1,16 @@
 import path from 'path';
-import merge from 'deepmerge';
 
 import { RENATIVE_CONFIG_NAME, RENATIVE_CONFIG_PRIVATE_NAME, RENATIVE_CONFIG_LOCAL_NAME } from '../constants';
-import {
-    sanitizeDynamicRefs,
-    sanitizeDynamicProps,
-    mergeObjects,
-    fsExistsSync,
-    fsReaddirSync,
-    fsLstatSync,
-    loadFile,
-    formatBytes,
-    mkdirSync,
-    writeFileSync,
-    getRealPath,
-    readObjectSync,
-} from '../system/fs';
-import { getConfigProp } from '../common';
-import { getWorkspaceDirPath } from '../workspaces';
-import { chalk, logTask, logWarning, logDebug } from '../logger';
+import { mergeObjects, fsExistsSync, fsReaddirSync, getRealPath, readObjectSync, loadFile } from '../system/fs';
+import { logTask, logWarning, logDebug } from '../logger';
 import { doResolve } from '../system/resolve';
 import { RnvContextFileObj, RnvContextPathObj, RnvContext, RnvContextFileKey } from '../context/types';
 import { generateRnvConfigPathObj } from '../context/defaults';
 import { generateContextPaths } from '../context';
-import { getContext } from '../context/provider';
-import { RnvPlatform } from '../types';
-// import { loadPluginTemplates } from '../pluginManager';
-
-const IGNORE_FOLDERS = ['.git'];
-
-export const checkIsRenativeProject = (c: RnvContext) =>
-    new Promise((resolve, reject) => {
-        if (!c.paths.project.configExists) {
-            return reject(
-                `This directory is not ReNative project. Project config ${chalk().white(
-                    c.paths.project.config
-                )} is missing!. You can create new project with ${chalk().white('rnv new')}`
-            );
-        }
-
-        return resolve(true);
-    });
-
-const _arrayMergeOverride = (_destinationArray: Array<string>, sourceArray: Array<string>) => sourceArray;
-
-const getEnginesPluginDelta = (c: RnvContext) => {
-    logDebug('getEnginesPluginDelta');
-
-    if (!c.buildConfig) return;
-
-    const enginePlugins: Record<string, any> = {};
-    const missingEnginePlugins: Record<string, any> = {};
-
-    const engineConfig = c.platform ? c.runtime.enginesByPlatform[c.platform]?.config : undefined;
-    if (engineConfig?.plugins) {
-        const ePlugins = Object.keys(engineConfig.plugins);
-
-        if (ePlugins?.length) {
-            ePlugins.forEach((pluginKey) => {
-                if (!c.files?.project?.config?.[pluginKey]) {
-                    missingEnginePlugins[pluginKey] = engineConfig.plugins?.[pluginKey];
-                }
-                enginePlugins[pluginKey] = engineConfig.plugins?.[pluginKey];
-            });
-        }
-    }
-    c.runtime.missingEnginePlugins = missingEnginePlugins;
-    return enginePlugins;
-};
-
-export const writeRenativeConfigFile = (c: RnvContext, configPath: string | undefined, configData: string | object) => {
-    logDebug(`writeRenativeConfigFile:${configPath}`);
-    writeFileSync(configPath, configData);
-    generateBuildConfig(c);
-};
-
-export const generateBuildConfig = (_c?: RnvContext) => {
-    logDebug('generateBuildConfig');
-
-    const c = _c || getContext();
-
-    const mergeOrder = [
-        c.paths.defaultWorkspace.config,
-        c.paths.rnv.projectTemplates.config,
-        c.paths.rnv.pluginTemplates.config,
-        // c.paths.rnv.platformTemplates.config,
-        c.paths.workspace.config,
-        c.paths.workspace.configPrivate,
-        c.paths.workspace.configLocal,
-        c.paths.workspace.project.config,
-        c.paths.workspace.project.configPrivate,
-        c.paths.workspace.project.configLocal,
-        ...c.paths.workspace.appConfig.configs,
-        ...c.paths.workspace.appConfig.configsPrivate,
-        ...c.paths.workspace.appConfig.configsLocal,
-        c.paths.project.config,
-        c.paths.project.configPrivate,
-        c.paths.project.configLocal,
-        ...c.paths.appConfig.configs,
-        ...c.paths.appConfig.configsPrivate,
-        ...c.paths.appConfig.configsLocal,
-    ];
-    const cleanPaths = mergeOrder.filter((v) => v);
-    const existsPaths = cleanPaths.filter((v) => {
-        const exists = fsExistsSync(v);
-        if (exists) {
-            logDebug(`Merged: ${v}`);
-        } else {
-            // console.log(chalk().red(v));
-        }
-        return exists;
-    });
-
-    const pluginTemplates: Record<string, any> = {};
-    if (c.files.rnv.pluginTemplates.configs) {
-        Object.keys(c.files.rnv.pluginTemplates.configs).forEach((v) => {
-            const plgs = c.files.rnv.pluginTemplates.configs[v];
-            pluginTemplates[v] = plgs;
-        });
-    }
-
-    const extraPlugins = getEnginesPluginDelta(c);
-
-    const mergeFiles = [
-        c.files.defaultWorkspace.config,
-        c.files.rnv.projectTemplates.config,
-        { plugins: extraPlugins },
-        // { pluginTemplates },
-        c.files.workspace.config,
-        c.files.workspace.configPrivate,
-        c.files.workspace.configLocal,
-        c.files.workspace.project.config,
-        c.files.workspace.project.configPrivate,
-        c.files.workspace.project.configLocal,
-        ...c.files.workspace.appConfig.configs,
-        ...c.files.workspace.appConfig.configsPrivate,
-        ...c.files.workspace.appConfig.configsLocal,
-        c.files.project.config,
-        c.files.project.configPrivate,
-        c.files.project.configLocal,
-        ...c.files.appConfig.configs,
-        ...c.files.appConfig.configsPrivate,
-        ...c.files.appConfig.configsLocal,
-    ];
-
-    // mergeFiles.forEach((mergeFile, i) => {
-    //     console.log(`MERGEDIAGNOSTICS ${i}`, Object.keys(mergeFile?.plugins || {}));
-    // });
-
-    const meta = [
-        {
-            _meta: {
-                generated: new Date().getTime(),
-                mergedConfigs: existsPaths,
-            },
-        },
-    ];
-    const existsFiles = mergeFiles.filter((v) => v);
-
-    logDebug(
-        `generateBuildConfig:mergeOrder.length:${mergeOrder.length},cleanPaths.length:${cleanPaths.length},existsPaths.length:${existsPaths.length},existsFiles.length:${existsFiles.length}`
-    );
-
-    let out: any = merge.all([...meta, ...existsFiles], {
-        arrayMerge: _arrayMergeOverride,
-    });
-    out = merge({}, out);
-    out.pluginTemplates = pluginTemplates;
-
-    c.buildConfig = sanitizeDynamicRefs(c, out);
-    const propConfig = {
-        files: c.files,
-        runtimeProps: c.runtime,
-        props: c.buildConfig._refs,
-        configProps: c.configPropsInjects,
-    };
-    c.buildConfig = sanitizeDynamicProps(c.buildConfig, propConfig);
-
-    logDebug('BUILD_CONFIG', Object.keys(c.buildConfig));
-
-    if (c.runtime.appId) {
-        c.paths.project.builds.config = path.join(c.paths.project.builds.dir, `${c.runtime.appId}_${c.platform}.json`);
-
-        logDebug(`generateBuildConfig: will sanitize file at: ${c.paths.project.builds.config}`);
-
-        if (c.paths.project.builds.dir) {
-            if (!fsExistsSync(c.paths.project.builds.dir)) {
-                mkdirSync(c.paths.project.builds.dir);
-            }
-
-            const result = writeFileSync(c.paths.project.builds.config, c.buildConfig);
-            if (result) {
-                const size = formatBytes(Buffer.byteLength(result || '', 'utf8'));
-                logTask(chalk().grey('generateBuildConfig'), `size:${size}`);
-            } else {
-                logDebug(`generateBuildConfig NOT SAVED: ${c.paths.project.builds.config}`);
-            }
-        } else {
-            logWarning('Cannot save buildConfig as c.paths.project.builds.dir is not defined');
-        }
-    }
-};
+import { generateBuildConfig } from './buildConfig';
+import { generateLocalConfig } from './configLocal';
+import { getWorkspaceDirPath } from './workspaces';
+import { generatePlatformTemplatePaths } from './configProject';
 
 export const loadFileExtended = (
     c: RnvContext,
@@ -367,156 +177,6 @@ const _loadConfigFiles = (
     return result;
 };
 
-export const generateRuntimeConfig = async (c: RnvContext) => {
-    logTask('generateRuntimeConfig');
-    // c.assetConfig = {
-    //     common: c.buildConfig.common,
-    //     runtime: c.buildConfig.runtime
-    // };
-    c.assetConfig = mergeObjects(c, c.assetConfig, c.buildConfig.runtime || {});
-    c.assetConfig = mergeObjects(c, c.assetConfig, c.buildConfig.common?.runtime || {});
-    c.assetConfig = mergeObjects(
-        c,
-        c.assetConfig,
-        c.platform ? c.buildConfig.platforms?.[c.platform]?.runtime || {} : {}
-    );
-    c.assetConfig = mergeObjects(c, c.assetConfig, getConfigProp(c, c.platform, 'runtime') || {});
-
-    if (fsExistsSync(c.paths.project.assets.dir)) {
-        const sanitizedConfig = sanitizeDynamicProps(c.assetConfig, {
-            files: c.files,
-            runtimeProps: c.runtime,
-            props: {},
-            configProps: c.configPropsInjects,
-        });
-        writeFileSync(c.paths.project.assets.config, sanitizedConfig);
-        c.files.project.assets.config = sanitizedConfig;
-    }
-    return true;
-};
-
-export const generateLocalConfig = (c: RnvContext, resetAppId?: boolean) => {
-    logTask('generateLocalConfig', `resetAppId:${!!resetAppId}`);
-    const configLocal = c.files.project.configLocal || {};
-    configLocal._meta = configLocal._meta || {};
-    if (resetAppId) {
-        delete configLocal._meta.currentAppConfigId;
-    } else {
-        configLocal._meta.currentAppConfigId = c.runtime.appId;
-    }
-    c.files.project.configLocal = configLocal;
-    writeFileSync(c.paths.project.configLocal, configLocal);
-};
-
-const _generatePlatformTemplatePaths = (c: RnvContext) => {
-    logTask('_generatePlatformTemplatePaths');
-    if (!c.buildConfig.paths) {
-        logWarning(`You're missing paths object in your ${chalk().red(c.paths.project.config)}`);
-        c.buildConfig.paths = {
-            appConfigsDirs: [],
-            platformTemplatesDirs: {},
-        };
-    }
-
-    const pt = c.buildConfig.paths?.platformTemplatesDirs || {};
-    const result: Record<string, string> = {};
-
-    if (c.buildConfig.defaults) {
-        c.buildConfig.defaults?.supportedPlatforms?.forEach((platform: RnvPlatform) => {
-            if (platform) {
-                const engine = c.runtime.enginesByPlatform[platform];
-                if (engine) {
-                    const originalPath = engine.originalTemplatePlatformsDir;
-
-                    if (originalPath) {
-                        if (!pt[platform]) {
-                            const pt1 = getRealPath(c, originalPath, 'platformTemplatesDir', originalPath);
-                            if (pt1) result[platform] = pt1;
-                        } else {
-                            const pt2 = getRealPath(c, pt[platform], 'platformTemplatesDir', originalPath);
-                            if (pt2) result[platform] = pt2;
-                        }
-                    } else {
-                        logWarning(
-                            `Platform ${chalk().red(platform)} not supported by any registered engine. SKIPPING...`
-                        );
-                    }
-                }
-            }
-        });
-    } else {
-        logWarning(`Your renative.json is missing property: ${chalk().red('defaults.supportedPlatforms')} `);
-    }
-
-    return result;
-};
-
-export const listAppConfigsFoldersSync = (c: RnvContext, ignoreHiddenConfigs: boolean, appConfigsDirPath?: string) => {
-    logTask('listAppConfigsFoldersSync', `ignoreHiddenConfigs:${!!ignoreHiddenConfigs}`);
-
-    if (!c.paths?.project) return [];
-
-    const dirPath = appConfigsDirPath || c.paths.project.appConfigsDir;
-
-    if (!fsExistsSync(dirPath)) return [];
-    const appConfigsDirs: Array<string> = [];
-    fsReaddirSync(dirPath).forEach((dir) => {
-        const appConfigDir = path.join(dirPath, dir);
-        if (!IGNORE_FOLDERS.includes(dir) && fsLstatSync(appConfigDir).isDirectory()) {
-            if (ignoreHiddenConfigs) {
-                const appConfig = path.join(appConfigDir, RENATIVE_CONFIG_NAME);
-                if (fsExistsSync(appConfig)) {
-                    try {
-                        const config = readObjectSync(appConfig);
-                        if (config?.hidden !== true) {
-                            appConfigsDirs.push(dir);
-                        }
-                    } catch (e) {
-                        logWarning(`_listAppConfigsFoldersSync: ${e}`);
-                    }
-                }
-            } else {
-                appConfigsDirs.push(dir);
-            }
-        }
-    });
-    return appConfigsDirs;
-};
-
-export const loadWorkspacesSync = () => {
-    const c = getContext();
-    // CHECK WORKSPACES
-    if (fsExistsSync(c.paths.rnv.configWorkspaces)) {
-        logDebug(`${c.paths.rnv.configWorkspaces} file exists!`);
-        c.files.rnv.configWorkspaces = readObjectSync(c.paths.rnv.configWorkspaces);
-
-        if (!c.files.rnv.configWorkspaces) c.files.rnv.configWorkspaces = {};
-
-        if (!c.files.rnv.configWorkspaces?.workspaces) {
-            c.files.rnv.configWorkspaces.workspaces = {};
-        }
-        if (Object.keys(c.files.rnv.configWorkspaces.workspaces).length === 0) {
-            logWarning(`No workspace found in ${c.paths.rnv.configWorkspaces}. Creating default rnv one for you`);
-            c.files.rnv.configWorkspaces.workspaces = {
-                rnv: {
-                    path: c.paths.workspace.dir,
-                },
-            };
-            writeFileSync(c.paths.rnv.configWorkspaces, c.files.rnv.configWorkspaces);
-        }
-    } else {
-        logWarning(`Cannot find ${c.paths.rnv.configWorkspaces}. creating one..`);
-        c.files.rnv.configWorkspaces = {
-            workspaces: {
-                rnv: {
-                    path: c.paths.workspace.dir,
-                },
-            },
-        };
-        writeFileSync(c.paths.rnv.configWorkspaces, c.files.rnv.configWorkspaces);
-    }
-};
-
 export const parseRenativeConfigs = async (c: RnvContext) => {
     logTask('parseRenativeConfigs');
     // LOAD ./package.json
@@ -570,7 +230,7 @@ export const parseRenativeConfigs = async (c: RnvContext) => {
 
     c.paths.workspace.project.appConfigBase.dir_LEGACY = path.join(c.paths.workspace.project.dir, 'projectConfig');
 
-    c.paths.project.platformTemplatesDirs = _generatePlatformTemplatePaths(c);
+    generatePlatformTemplatePaths(c);
 
     if (c.runtime.appId) {
         if (!c.files.appConfig.config) {
