+++ packages/core/src/index.ts
@@ -9,18 +9,19 @@ export * from './engines';
 export * from './platforms';
 export * from './plugins';
 export * from './projects';
-export * from './workspaces';
 export * from './buildHooks';
 export * from './migrator';
 export * from './configs';
-export * from './configs/schemeParser';
-export * from './configs/packageParser';
+export * from './configs/buildSchemes';
+export * from './configs/buildConfig';
+export * from './configs/workspaces';
+export * from './projects/dependencyManager';
 export * from './schema/zod';
 export * from './runtime';
 export * from './templates';
 export * from './integrations';
 export * from './tasks';
-export * from './npm';
+export * from './projects/npm';
 export * from './git';
 export * from './utils/is';
 export * from './system/exec';
